move legacy http remote state to a backend
