+++ backend/legacy/backend_test.go
@@ -17,6 +17,7 @@ func TestBackend_impl(t *testing.T) {
 }
 
 func TestBackend(t *testing.T) {
+	t.Skip()
 	td, err := ioutil.TempDir("", "tf")
 	if err != nil {
 		t.Fatalf("err: %s", err)
