+++ state/remote/remote.go
@@ -46,6 +46,4 @@ func NewClient(t string, conf map[string]string) (Client, error) {
 // NewClient.
 var BuiltinClients = map[string]Factory{
 	"artifactory": artifactoryFactory,
-	"http":        httpFactory,
-	"local":       fileFactory,
 }
